decouple labels on relationships from relationship view component
