+++ packages/tc-ui/primitives/relationship/lib/view-relationship.jsx
@@ -2,77 +2,14 @@ const { h, Fragment } = require('preact');
 const { getType } = require('@financial-times/tc-schema-sdk');
 const { LinkToRecord } = require('../../../lib/helpers');
 
-const oLabelsModifiersMap = {
-	platinum: 'tier-platinum',
-	gold: 'tier-gold',
-	silver: 'tier-silver',
-	bronze: 'tier-bronze',
-	unsupported: 'support-dead',
-	undefined: 'unknown',
-	null: 'unknown',
-	unknown: 'unknown',
-	inactive: 'support-dead',
-	preproduction: 'support-experimental',
-	production: 'support-active',
-	deprecated: 'support-deprecated',
-	decommissioned: 'support-dead',
-	incubate: 'support-experimental',
-	sustain: 'support-active',
-	grow: 'support-active',
-	sunset: 'support-deprecated',
-};
+let RelationshipAnnotator;
 
-const Label = ({ value, fallbackText, id }) => (
-	<span
-		id={id}
-		className={`o-layout--unstyled-element o-labels o-labels--${
-			oLabelsModifiersMap[value ? value.toLowerCase() : 'unknown']
-		}`}
-	>
-		{value || fallbackText}
-	</span>
-);
-
-const getLabelWithFallback = fallbackText => props =>
-	Label({ fallbackText, ...props });
-
-const LifecycleStage = getLabelWithFallback('Unknown lifecycle stage');
-
-const ServiceTier = getLabelWithFallback('Unknown service tier');
-
-const IsActiveLabel = ({ isActive }) =>
-	isActive ? null : (
-		<span
-			className={`o-labels o-labels--${
-				oLabelsModifiersMap[isActive === false ? 'inactive' : 'unknown']
-			}`}
-		>
-			{isActive === false ? 'Inactive' : 'May not be active'}
-		</span>
-	);
-
-const OneRelationship = ({ type, value = {}, id, annotate = true }) =>
+const OneRelationship = ({ type, value = {}, id }) =>
 	type && value.code ? (
 		<Fragment>
 			<LinkToRecord id={id} type={type} value={value} />
-			{annotate ? (
-				<Fragment>
-					{type === 'System' ? (
-						<ServiceTier value={value.serviceTier} />
-					) : null}
-					{type === 'System' ? (
-						<LifecycleStage value={value.lifecycleStage} />
-					) : null}
-					{type === 'Person' ? (
-						<IsActiveLabel isActive={value.isActive} />
-					) : null}
-					{type === 'Team' ? (
-						<IsActiveLabel isActive={value.isActive} />
-					) : null}
-					{type === 'Repository' ? (
-						<IsActiveLabel isActive={!value.isArchived} />
-					) : null}
-				</Fragment>
+			{RelationshipAnnotator ? (
+				<RelationshipAnnotator value={value} type={type} />
 			) : null}
 		</Fragment>
 	) : (
@@ -109,4 +46,9 @@ const ViewRelationship = ({ value, type, id, hasMany }) => {
 	);
 };
 
-module.exports = { ViewRelationship };
+module.exports = {
+	ViewRelationship,
+	setRelationshipAnnotator: component => {
+		RelationshipAnnotator = component;
+	},
+};
