+++ packages/tc-ui/primitives/relationship/server.jsx
@@ -3,7 +3,10 @@ const { h } = require('preact');
 const { FieldTitle } = require('../../lib/edit-helpers');
 
 const { RelationshipPicker } = require('./lib/relationship-picker');
-const { ViewRelationship } = require('./lib/view-relationship');
+const {
+	ViewRelationship,
+	setRelationshipAnnotator,
+} = require('./lib/view-relationship');
 
 const EditRelationship = props => (
 	// eslint-disable-next-line jsx-a11y/label-has-associated-control, jsx-a11y/label-has-for
@@ -19,4 +22,5 @@ module.exports = {
 	parser: value => (value ? JSON.parse(value) : null),
 	hasValue: (value, { hasMany }) =>
 		hasMany ? value && value.length : !!value,
+	setRelationshipAnnotator,
 };
